feat(package.json): add new dependencies for @lucia-auth/adapter-drizzle, lucia, and oslo
The @lucia-auth/adapter-drizzle dependency is added to provide support for using the Drizzle adapter in the Lucia authentication library.

The lucia dependency is added to provide support for using the Lucia library in the project.

The oslo dependency is added to provide support for using the Oslo library in the project.